fix: purchase options not working
